+++ src/contracts/handlers/default.ts
@@ -7,11 +7,8 @@ import {
     PathContext,
     PathSelection,
 } from "../types";
-import {
-    isCsvSpendable,
-    sequenceToTimelock,
-    timelockToSequence,
-} from "./helpers";
+import { isCsvSpendable } from "./helpers";
+import { sequenceToTimelock, timelockToSequence } from "../../utils/timelock";
 import {
     normalizeToDescriptor,
     extractPubKey,
